check for sha1

+++ python/mxnet/gluon/utils.py
@@ -294,9 +294,10 @@ def download(url, path=None, overwrite=False, sha1_hash=None, retries=5, verify_
                     for chunk in r.iter_content(chunk_size=1024):
                         if chunk: # filter out keep-alive new chunks
                             f.write(chunk)
-                # if the target file exists(created by other processes),
+                # if the target file exists(created by other processes)
+                # and have the same hash with target file
                 # delete the temporary file
-                if os.path.exists(fname):
+                if os.path.exists(fname) and sha1_hash and check_sha1(fname, sha1_hash):
                     os.remove('{}.{}'.format(fname, random_uuid))
                     warnings.warn(
                         'File {} exists in file system so the downloaded file is deleted'.format(fname))
